python-task-manager/requirements.txt
DJango>=4.0
psycopg2-binary
djangorestframework
django-cors-headers
